requirements.txt
pip == 21.1
setuptools == 54.2.0
flwr==0.16.0
pandas==1.1.5
python-socketio==5.1.0
mlflow==1.24.0
dynaconf==3.1.4
git+https://github.com/kosmos-industrie40/kosmos-federated-learning-resources.git@release
